Add routing for basic and interactive chart components
