+++ ui/src/app/app-routing.module.ts
@@ -5,19 +5,31 @@ import { ArtistComponent } from './artist/artist.component';
 import { ChartDisplayComponent } from './chart-display/chart-display.component';
 import { ChartEditComponent } from './chart-edit/chart-edit.component';
 import { ChartSelectComponent } from './chart-select/chart-select.component';
-import { CreateChartComponent } from './create-chart/create-chart.component';
+import { CreateBasicChartComponent } from './create-chart-basic/create-chart-basic.component';
+import { CreateInteractiveChartComponent } from './create-chart-interactive/create-chart-interactive.component';
 import { SeriesSelectComponent } from './series-select/series-select.component';
 import { SongDisplayComponent } from './song-display/song-display.component';
 import { SongComponent } from './song/song.component';
 import { TotalsComponent } from './totals/totals.component';
 import { HomeComponent } from './home/home.component';
+import { CreateChartComponent } from './create-chart/create-chart.component';
 
 const routes: Routes = [
   // Can't have 'children' unless we want the parent displayed also!
   { path: '', component: HomeComponent },
   { path: 'series', component: SeriesSelectComponent },
   { path: 'series/:series/chart', component: ChartSelectComponent },
-  { path: 'series/:series/chart/new', component: CreateChartComponent },
+  { path: 'series/:series/chart/new',
+     component: CreateChartComponent,
+     children: [{
+      path: 'basic',
+      component: CreateBasicChartComponent,
+     },
+     {
+      path: 'interactive',
+      component: CreateInteractiveChartComponent,
+     }]
+  },
   { path: 'series/:series/chart/:name', component: ChartDisplayComponent },
   { path: 'series/:series/chart/:name/edit', component: ChartEditComponent },
   { path: 'song', component: SongComponent },
