+++ ui/src/app/create-chart-basic/create-chart-basic.component.css
@@ -1,0 +1,31 @@
+.input-option {
+    padding-bottom: 1rem;
+}
+
+#chart-text {
+    height: 75rem;
+    width: 25rem
+}
+
+.subtext {
+    color: gray;
+    font-size: small;
+}
+
+.submit-chart {
+    float: right;
+}
+
+.title-submit-table {
+    width: 100%
+}
+
+.centre {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    margin-top: -50px;
+    margin-left: -50px;
+    width: 10rem;
+    height: 10rem;
+}
