+++ ui/src/app/create-chart-basic/create-chart-basic.component.spec.ts
@@ -1,0 +1,28 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { RouterTestingModule } from '@angular/router/testing';
+
+import { CreateBasicChartComponent } from './create-chart-basic.component';
+
+describe('CreateBasicChartComponent', () => {
+  let component: CreateBasicChartComponent;
+  let fixture: ComponentFixture<CreateBasicChartComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [ HttpClientTestingModule, RouterTestingModule ],
+      declarations: [ CreateBasicChartComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(CreateBasicChartComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
