+++ ui/src/app/create-chart-interactive/create-chart-interactive.component.html
@@ -1,0 +1,1 @@
+<p>Coming soon!</p>
