+++ ui/src/app/create-chart-interactive/create-chart-interactive.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-create-chart-interactive',
+  templateUrl: './create-chart-interactive.component.html',
+  styleUrls: ['./create-chart-interactive.component.css']
+})
+export class CreateInteractiveChartComponent {
+
+}
