+++ ui/src/app/create-chart/create-chart.component.spec.ts
@@ -1,28 +1,59 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { Component } from '@angular/core';
+import { fakeAsync, TestBed, tick } from '@angular/core/testing';
+import { Routes } from '@angular/router';
 import { RouterTestingModule } from '@angular/router/testing';
-
 import { CreateChartComponent } from './create-chart.component';
 
-describe('CreateChartComponent', () => {
-  let component: CreateChartComponent;
-  let fixture: ComponentFixture<CreateChartComponent>;
+@Component({
+  template: `Basic`,
+})
+class TestBasicComponent {}
 
+@Component({
+  template: `Interactive`,
+})
+class TestInteractiveComponent {}
+
+
+export const routes: Routes = [
+  { path: 'basic', component: TestBasicComponent },
+  { path: 'interactive', component: TestInteractiveComponent },
+];
+
+describe('CreateChartComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [ HttpClientTestingModule, RouterTestingModule ],
-      declarations: [ CreateChartComponent ]
-    })
-    .compileComponents();
+      imports: [RouterTestingModule.withRoutes(routes)],
+      declarations: [
+        CreateChartComponent,
+      ],
+    }).compileComponents();
   });
 
-  beforeEach(() => {
-    fixture = TestBed.createComponent(CreateChartComponent);
-    component = fixture.componentInstance;
+  it('should create the app', () => {
+    const fixture = TestBed.createComponent(CreateChartComponent);
+    const app = fixture.componentInstance;
+    expect(app).toBeTruthy();
+  });
+
+  it(`should have as title 'create-chart'`, () => {
+    const fixture = TestBed.createComponent(CreateChartComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual('create-chart');
+  });
+
+  it('should route correctly', fakeAsync(() => {
+    const fixture = TestBed.createComponent(CreateChartComponent);
     fixture.detectChanges();
-  });
+    const compiled = fixture.nativeElement;
+    const buttons = compiled.querySelectorAll('a')
+    const pages = ['Basic', 'Interactive']
 
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
+    buttons.forEach((button, index) => {
+      button.click()
+      tick(1)
+      const component = fixture.nativeElement.querySelector('ng-component')
+      expect(component.textContent).withContext(`route to ${pages[index]}`).toContain(pages[index])
+    })
+  }));
 });
